requirements.txt
python-telegram-bot==21.11.1
Requests==2.32.3
